cars/requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
Django==3.0.2
django-allauth==0.41.0
django-colorfield==0.3.0
django-countries==6.1
django-crispy-forms==1.9.0
idna==2.9
oauthlib==3.1.0
Pillow==7.0.0
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
sqlparse==0.3.0
stripe==2.45.0
urllib3==1.25.8
virtualenv==16.7.5
